feat: restructure swapper module
