+++ lib/core/src/model.rs
@@ -18,7 +18,7 @@ use sdk_common::prelude::*;
 use serde::{Deserialize, Serialize};
 use strum_macros::{Display, EnumString};
 
-use crate::error::{PaymentError, SdkResult};
+use crate::error::{PaymentError, SdkError, SdkResult};
 use crate::receive_swap::{
     DEFAULT_ZERO_CONF_MAX_SAT, DEFAULT_ZERO_CONF_MIN_FEE_RATE_MAINNET,
     DEFAULT_ZERO_CONF_MIN_FEE_RATE_TESTNET,
@@ -166,6 +166,15 @@ impl From<LiquidNetwork> for sdk_common::prelude::Network {
 }
 
 impl From<LiquidNetwork> for sdk_common::bitcoin::Network {
+    fn from(value: LiquidNetwork) -> Self {
+        match value {
+            LiquidNetwork::Mainnet => Self::Bitcoin,
+            LiquidNetwork::Testnet => Self::Testnet,
+        }
+    }
+}
+
+impl From<LiquidNetwork> for boltz_client::bitcoin::Network {
     fn from(value: LiquidNetwork) -> Self {
         match value {
             LiquidNetwork::Mainnet => Self::Bitcoin,
@@ -661,7 +670,7 @@ pub(crate) struct SendSwap {
     pub(crate) refund_private_key: String,
 }
 impl SendSwap {
-    pub(crate) fn get_refund_keypair(&self) -> Result<Keypair, PaymentError> {
+    pub(crate) fn get_refund_keypair(&self) -> Result<Keypair, SdkError> {
         utils::decode_keypair(&self.refund_private_key).map_err(Into::into)
     }
 
@@ -688,12 +697,12 @@ impl SendSwap {
         Ok(res)
     }
 
-    pub(crate) fn get_swap_script(&self) -> Result<LBtcSwapScript, PaymentError> {
+    pub(crate) fn get_swap_script(&self) -> Result<LBtcSwapScript, SdkError> {
         LBtcSwapScript::submarine_from_swap_resp(
             &self.get_boltz_create_response()?,
             self.get_refund_keypair()?.public_key().into(),
         )
-        .map_err(|e| PaymentError::Generic {
+        .map_err(|e| SdkError::Generic {
             err: format!(
                 "Failed to create swap script for Send Swap {}: {e:?}",
                 self.id
@@ -1377,6 +1386,48 @@ pub struct LnUrlPaySuccessData {
 pub enum Transaction {
     Liquid(boltz_client::elements::Transaction),
     Bitcoin(boltz_client::bitcoin::Transaction),
+}
+
+#[derive(Debug, Clone)]
+pub enum Utxo {
+    Liquid(
+        (
+            boltz_client::elements::OutPoint,
+            boltz_client::elements::TxOut,
+        ),
+    ),
+    Bitcoin(
+        (
+            boltz_client::bitcoin::OutPoint,
+            boltz_client::bitcoin::TxOut,
+        ),
+    ),
+}
+
+impl Utxo {
+    pub(crate) fn as_bitcoin(
+        &self,
+    ) -> Option<&(
+        boltz_client::bitcoin::OutPoint,
+        boltz_client::bitcoin::TxOut,
+    )> {
+        match self {
+            Utxo::Liquid(_) => None,
+            Utxo::Bitcoin(utxo) => Some(utxo),
+        }
+    }
+
+    pub(crate) fn as_liquid(
+        &self,
+    ) -> Option<&(
+        boltz_client::elements::OutPoint,
+        boltz_client::elements::TxOut,
+    )> {
+        match self {
+            Utxo::Bitcoin(_) => None,
+            Utxo::Liquid(utxo) => Some(utxo),
+        }
+    }
 }
 
 #[macro_export]
